Cleaned up for Run 2 (M7) (TrigBjetMonitoring-00-01-12)
2014-12-03 Elemer Nagy <Elemer.Nagy AT cern.ch>
	* Clean up for Run 2 (M7)
	* TrigBjetMonitoring-00-01-12 tagged

+++ Trigger/TrigMonitoring/TrigBjetMonitoring/src/TuningLikelihood.cxx
@@ -19,7 +19,8 @@
 //** ----------------------------------------------------------------------------------------------------------------- **//
 
 
-TuningLikelihood::TuningLikelihood(const float *sizeLikelihood, const float *bLikelihood,
+//TuningLikelihood::TuningLikelihood(const float *sizeLikelihood, const float *bLikelihood, 
+TuningLikelihood::TuningLikelihood(const int *sizeLikelihood, const float *bLikelihood,
 				   const float *uLikelihood, const int dimensionLikelihood) :
   m_bLikelihood(&bLikelihood[0]),
   m_uLikelihood(&uLikelihood[0]),
